Isolate stock split planning in cgt_calc/stock_split_planning.py
Moves the first-pass share-reorganisation planning family out of
TransactionIngester into a module exposing plan_stock_splits and
source_account, with the planning functions taking the shared
CalculatorState explicitly. The planned SplitTransformations are
the contract both passes consume; value types and arithmetic stay
in stock_splits.py, spin-offs stay in ingestion, replay stays in
Matcher. No behavior change.
